refactor: rename diagram file name util
